+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/module/installer/bundle/GuiceyBootstrap.java
@@ -13,6 +13,7 @@ import ru.vyarus.dropwizard.guice.module.lifecycle.GuiceyLifecycleListener;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.function.Consumer;
 import java.util.function.Supplier;
 
 /**
@@ -223,5 +224,13 @@ public class GuiceyBootstrap implements GuiceyCommonRegistration<GuiceyBootstrap
     @Override
     public <K> K sharedStateOrFail(final Class<?> key, final String message, final Object... args) {
         return context.getSharedState().getOrFail(key, message, args);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public <V> void whenSharedStateReady(final Class<?> key, final Consumer<V> action) {
+        context.getSharedState().whenReady(key, action);
     }
 }