+++ dropwizard-guicey/src/test/groovy/ru/vyarus/dropwizard/guice/config/sharedstate/SharedHookStateTest.groovy
@@ -9,6 +9,7 @@ import ru.vyarus.dropwizard.guice.hook.GuiceyConfigurationHook
 import ru.vyarus.dropwizard.guice.module.context.SharedConfigurationState
 import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyBootstrap
 import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyBundle
+import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyEnvironment
 import ru.vyarus.dropwizard.guice.test.jupiter.TestGuiceyApp
 import spock.lang.Specification
 
@@ -24,10 +25,14 @@ class SharedHookStateTest extends Specification {
     @Inject
     Bootstrap bootstrap
 
-    def "Check hook access to shared memeory"() {
+    def "Check hook access to shared memory"() {
 
         expect:
         SharedConfigurationState.lookup(bootstrap.getApplication(), XHook).get() == "12"
+        SharedConfigurationState.lookup(bootstrap.getApplication(), Bundle).get() == "15"
+        SharedConfigurationState.lookup(bootstrap.getApplication(), SharedHookStateTest).get() == "20"
+        Bundle.called
+        Bundle.called2
     }
 
     static class App extends Application<Configuration> {
@@ -53,9 +58,21 @@ class SharedHookStateTest extends Specification {
     }
 
     static class Bundle implements GuiceyBundle {
+
+        static boolean called
+        static boolean called2
+
         @Override
         void initialize(GuiceyBootstrap bootstrap) {
+            bootstrap.shareState(SharedHookStateTest, "20")
+            bootstrap.whenSharedStateReady(Bundle, { assert it == "15"; called = true })
             assert bootstrap.sharedStateOrFail(XHook, "ugr") == "12"
+        }
+
+        @Override
+        void run(GuiceyEnvironment environment) throws Exception {
+            environment.whenSharedStateReady(Bundle, { assert it == "15"; called2 = true})
+            environment.shareState(Bundle, "15")
         }
     }
 }